requirements_test.txt
pytest
pytest-cov
pytest-homeassistant-custom-component
aiohttp
syrupy
fing-agent-api
